# apps/backend/src/generated/models/AuthSession.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthSession
 */
export interface AuthSession {
    /**
     * 
     * @type {string}
     * @memberof AuthSession
     */
    userId?: string;
}

/**
 * Check if a given object implements the AuthSession interface.
 */
export function instanceOfAuthSession(value: object): value is AuthSession {
    return true;
}

export function AuthSessionFromJSON(json: any): AuthSession {
    return AuthSessionFromJSONTyped(json, false);
}

export function AuthSessionFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthSession {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'] == null ? undefined : json['userId'],
    };
}

export function AuthSessionToJSON(value?: AuthSession | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'userId': value['userId'],
    };
}