# apps/backend/src/generated/models/CreateListingInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateListingInput
 */
export interface CreateListingInput {
    /**
     * 
     * @type {string}
     * @memberof CreateListingInput
     */
    title: string;
    /**
     * 
     * @type {number}
     * @memberof CreateListingInput
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof CreateListingInput
     */
    bedrooms: number;
}

/**
 * Check if a given object implements the CreateListingInput interface.
 */
export function instanceOfCreateListingInput(value: object): value is CreateListingInput {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    if (!('bedrooms' in value) || value['bedrooms'] === undefined) return false;
    return true;
}

export function CreateListingInputFromJSON(json: any): CreateListingInput {
    return CreateListingInputFromJSONTyped(json, false);
}

export function CreateListingInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateListingInput {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'price': json['price'],
        'bedrooms': json['bedrooms'],
    };
}

export function CreateListingInputToJSON(value?: CreateListingInput | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'price': value['price'],
        'bedrooms': value['bedrooms'],
    };
}