# apps/backend/src/generated/models/ListingFilter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListingFilter
 */
export interface ListingFilter {
    /**
     * 
     * @type {number}
     * @memberof ListingFilter
     */
    minPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof ListingFilter
     */
    maxPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof ListingFilter
     */
    bedrooms?: number;
    /**
     * 
     * @type {string}
     * @memberof ListingFilter
     */
    status?: ListingFilterStatusEnum;
}


/**
 * @export
 */
export const ListingFilterStatusEnum = {
    Draft: 'draft',
    Published: 'published',
    Archived: 'archived'
} as const;
export type ListingFilterStatusEnum = typeof ListingFilterStatusEnum[keyof typeof ListingFilterStatusEnum];


/**
 * Check if a given object implements the ListingFilter interface.
 */
export function instanceOfListingFilter(value: object): value is ListingFilter {
    return true;
}

export function ListingFilterFromJSON(json: any): ListingFilter {
    return ListingFilterFromJSONTyped(json, false);
}

export function ListingFilterFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListingFilter {
    if (json == null) {
        return json;
    }
    return {
        
        'minPrice': json['minPrice'] == null ? undefined : json['minPrice'],
        'maxPrice': json['maxPrice'] == null ? undefined : json['maxPrice'],
        'bedrooms': json['bedrooms'] == null ? undefined : json['bedrooms'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListingFilterToJSON(value?: ListingFilter | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'minPrice': value['minPrice'],
        'maxPrice': value['maxPrice'],
        'bedrooms': value['bedrooms'],
        'status': value['status'],
    };
}