# apps/backend/src/generated/models/ListingPage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PaginationMeta } from './PaginationMeta';
import {
    PaginationMetaFromJSON,
    PaginationMetaFromJSONTyped,
    PaginationMetaToJSON,
} from './PaginationMeta';
import type { Listing } from './Listing';
import {
    ListingFromJSON,
    ListingFromJSONTyped,
    ListingToJSON,
} from './Listing';

/**
 * 
 * @export
 * @interface ListingPage
 */
export interface ListingPage {
    /**
     * 
     * @type {Array<Listing>}
     * @memberof ListingPage
     */
    data?: Array<Listing>;
    /**
     * 
     * @type {PaginationMeta}
     * @memberof ListingPage
     */
    meta?: PaginationMeta;
}

/**
 * Check if a given object implements the ListingPage interface.
 */
export function instanceOfListingPage(value: object): value is ListingPage {
    return true;
}

export function ListingPageFromJSON(json: any): ListingPage {
    return ListingPageFromJSONTyped(json, false);
}

export function ListingPageFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListingPage {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(ListingFromJSON)),
        'meta': json['meta'] == null ? undefined : PaginationMetaFromJSON(json['meta']),
    };
}

export function ListingPageToJSON(value?: ListingPage | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(ListingToJSON)),
        'meta': PaginationMetaToJSON(value['meta']),
    };
}