# apps/backend/src/generated/models/UpdateListingInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateListingInput
 */
export interface UpdateListingInput {
    /**
     * 
     * @type {string}
     * @memberof UpdateListingInput
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateListingInput
     */
    bedrooms?: number;
}

/**
 * Check if a given object implements the UpdateListingInput interface.
 */
export function instanceOfUpdateListingInput(value: object): value is UpdateListingInput {
    return true;
}

export function UpdateListingInputFromJSON(json: any): UpdateListingInput {
    return UpdateListingInputFromJSONTyped(json, false);
}

export function UpdateListingInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateListingInput {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'bedrooms': json['bedrooms'] == null ? undefined : json['bedrooms'],
    };
}

export function UpdateListingInputToJSON(value?: UpdateListingInput | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'bedrooms': value['bedrooms'],
    };
}