# packages/types/src/generated/models/listing.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Real Estate API
 * Canonical contract for listing and authentication APIs.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Listing
 */
export interface Listing {
    /**
     * 
     * @type {string}
     * @memberof Listing
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Listing
     */
    'title'?: string;
    /**
     * 
     * @type {number}
     * @memberof Listing
     */
    'price'?: number;
    /**
     * 
     * @type {number}
     * @memberof Listing
     */
    'bedrooms'?: number;
    /**
     * 
     * @type {string}
     * @memberof Listing
     */
    'status'?: ListingStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Listing
     */
    'createdAt'?: string;
}

export const ListingStatusEnum = {
    Draft: 'draft',
    Published: 'published',
    Archived: 'archived'
} as const;

export type ListingStatusEnum = typeof ListingStatusEnum[keyof typeof ListingStatusEnum];