# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Real Estate API
  version: 1.0.0
  description: Canonical contract for listing and authentication APIs.
  contact:
    name: Real Estate Platform Team
servers:
  - url: http://localhost:3001
    description: Local dev
tags:
  - name: Listings
    description: Listing management endpoints.
  - name: Auth
    description: Authentication endpoints.
paths:
  /auth/signup:
    post:
      operationId: signUp
      summary: Sign up
      description: Registers a new user with email and password.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpInput'
      responses:
        '200':
          description: User registered and session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/signin:
    post:
      operationId: signIn
      summary: Sign in
      description: Authenticates a user and creates a session.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInInput'
      responses:
        '200':
          description: User authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/ValidationError'
  /auth/signout:
    post:
      operationId: signOut
      summary: Sign out
      description: Revokes the active session.
      tags: [Auth]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Signed out
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /auth/session:
    get:
      operationId: getSession
      summary: Get session
      description: Returns the active authenticated session.
      tags: [Auth]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/ValidationError'
  /listings:
    get:
      operationId: getListings
      summary: Get paginated listings
      description: Returns a paginated list of property listings with sorting and filter options.
      tags: [Listings]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: filter
          in: query
          style: deepObject
          explode: true
          schema:
            $ref: '#/components/schemas/ListingFilter'
      responses:
        '200':
          description: A page of listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPage'
    post:
      operationId: createListing
      summary: Create a new listing
      description: Creates a draft property listing.
      tags: [Listings]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingInput'
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '422':
          $ref: '#/components/responses/ValidationError'
  /listings/{id}/publish:
    post:
      operationId: publishListing
      summary: Publish a listing
      description: Publishes a draft listing if its domain invariants are satisfied.
      tags: [Listings]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Published listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '422':
          $ref: '#/components/responses/ValidationError'
  /listings/{id}:
    get:
      operationId: getListing
      summary: Get a single listing
      description: Returns a single listing by identifier.
      tags: [Listings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateListing
      summary: Update a listing (draft only)
      description: Updates mutable listing fields while the listing is still in draft status.
      tags: [Listings]
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListingInput'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '422':
          $ref: '#/components/responses/ValidationError'

components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [createdAt_asc, createdAt_desc, price_asc, price_desc]
        default: createdAt_desc

  schemas:
    SignUpInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    SignInInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthSession:
      type: object
      properties:
        userId:
          type: string

    PaginationMeta:
      type: object
      properties:
        totalCount:
          type: integer
        pageCount:
          type: integer
        currentPage:
          type: integer
        perPage:
          type: integer

    ListingFilter:
      type: object
      properties:
        minPrice:
          type: number
        maxPrice:
          type: number
        bedrooms:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]

    CreateListingInput:
      type: object
      required: [title, price, bedrooms]
      properties:
        title:
          type: string
          minLength: 1
        price:
          type: number
          exclusiveMinimum: 0
        bedrooms:
          type: integer
          minimum: 0

    UpdateListingInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        bedrooms:
          type: integer
          minimum: 0

    Listing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        price:
          type: number
        bedrooms:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]
        createdAt:
          type: string
          format: date-time

    ListingPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sAccessToken